Create LibStore.py
This is what my bot has picked up in the last month, I get an original list and I either black or white list them, this is the blacklist.  Somehow the act of running the bot has increased the number of advertisements, I hope it plays merry havoc with their machine learning data on me.